restructure for header only format.
