+++ include/models/linear/modelHelpers.hpp
