+++ include/tools/autoReg.hpp
