+++ include/tools/npTools.hpp
