--- models/src/testModels.cpp
@@ -1,5 +1,0 @@
-#include "RegressionModel.hpp"
-
-int main(int argc, char* argv[]) {
-    return 0;
-}
