--- src/main.cpp
@@ -1,9 +1,0 @@
-#include <iostream>
-
-#include <ADFT.hpp>
-
-int main(int argc, char *argv[]) {
-
-
-    return 0;
-}
